Refactor AMA submit review
