+++ client/app/intake/actions/appeal.js
@@ -1,9 +1,6 @@
-import { ACTIONS, ENDPOINT_NAMES } from '../constants';
-import ApiUtil from '../../util/ApiUtil';
-import { formatDateStringForApi } from '../../util/DateUtil';
-import _ from 'lodash';
+import { ACTIONS } from '../constants';
 
-const analytics = true;
+// const analytics = true;
 
 export const setDocketType = (docketType) => ({
   type: ACTIONS.SET_DOCKET_TYPE,
