+++ client/app/intake/pages/supplementalClaim/review.jsx
@@ -6,7 +6,7 @@ import CancelButton from '../../components/CancelButton';
 import { Redirect } from 'react-router-dom';
 import Button from '../../../components/Button';
 import SelectClaimant from '../../components/SelectClaimant';
-import { submitReview } from '../../actions/supplementalClaim';
+import { submitReview } from '../../actions/ama';
 import { setReceiptDate, setClaimantNotVeteran, setClaimant } from '../../actions/common';
 import { REQUEST_STATE, PAGE_PATHS, INTAKE_STATES } from '../../constants';
 import { getIntakeStatus } from '../../selectors';
@@ -60,7 +60,7 @@ const SelectClaimantConnected = connect(
 
 class ReviewNextButton extends React.PureComponent {
   handleClick = () => {
-    this.props.submitReview(this.props.intakeId, this.props.supplementalClaim).then(
+    this.props.submitReview(this.props.intakeId, this.props.supplementalClaim, 'supplementalClaim').then(
       () => this.props.history.push('/finish')
     );
   };
