+++ client/app/intake/util/index.js
@@ -1,5 +1,6 @@
 import _ from 'lodash';
 import { REVIEW_OPTIONS } from '../constants';
+import { formatDateStringForApi } from '../../util/DateUtil';
 
 export const getAppealDocketError = (responseErrorCodes) => (
   (_.get(responseErrorCodes.appeal_docket, 0) === 'blank') && 'Please select an option.'
@@ -81,22 +82,27 @@ export const formatRatingData = (intakeState) => {
 
 export const prepareReviewData = (intakeData, intakeType) => {
   switch (intakeType) {
-  case "appeal":
+  case 'appeal':
     return {
       docket_type: intakeData.docketType,
       receipt_date: formatDateStringForApi(intakeData.receiptDate),
       claimant: intakeData.claimant
     };
-  case "supplementalClaim":
+  case 'supplementalClaim':
     return {
-        receipt_date: formatDateStringForApi(intakeData.receiptDate),
-        claimant: intakeData.claimant
-      };
-  case "higherLevelReview":
+      receipt_date: formatDateStringForApi(intakeData.receiptDate),
+      claimant: intakeData.claimant
+    };
+  case 'higherLevelReview':
     return {
       informal_conference: intakeData.informalConference,
       same_office: intakeData.sameOffice,
       receipt_date: formatDateStringForApi(intakeData.receiptDate),
       claimant: intakeData.claimant
     };
+  default:
+    return {
+      receipt_date: formatDateStringForApi(intakeData.receiptDate)
+    };
+  }
 };
